C12 - Added the ability to query all animals for animals by applying DTO.

+++ Src/Core/VirtualPaws.Application/Interfaces/Repository/Entities/IGenericEntityRepository.cs
@@ -6,7 +6,7 @@ namespace VirtualPaws.Application.Interfaces.Repository.Entities
     public interface IGenericEntityRepository<T> where T : BaseEntity
     {
         List<T> GetAll();
-        T GetById(int id);
+        T GetById(UInt16 id);
         void Delete(T entity);
         void Update(T entity);
         void Create(T entity);
